+++ web/src/Canvas.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { Box } from '@material-ui/core';
 
 // https://blog.koenvangilst.nl/react-hooks-with-canvas/
 // https://www.w3schools.com/Tags/canvas_putimagedata.asp
@@ -17,7 +18,11 @@ function Canvas(props) {
     }
   }, [imageData]);
 
-  return <canvas ref={canvasRef} width={width} height={height} />;
+  return (
+    <Box boxShadow={2}>
+      <canvas ref={canvasRef} width={width} height={height} />
+    </Box>
+  );
 }
 
 // TODO: get width and height from imagedata
