+++ web/src/Retropixels.jsx
@@ -1,11 +1,14 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import { Container, Grid, Slider, Typography } from '@material-ui/core';
+import { Box, Button, Container, Grid, Slider, Typography } from '@material-ui/core';
 import BlurLinearIcon from '@material-ui/icons/BlurLinear';
+import CloudDownloadIcon from '@material-ui/icons/CloudDownload';
+import { C64Writer } from 'retropixels-core';
+import { saveAs } from 'file-saver';
 import HiresCheckbox from './HiresCheckbox';
 import ProfileSelection from './ProfileSelection';
 import TargetImage from './TargetImage';
-import { parseFilename} from './Utilities';
+import { parseFilename } from './Utilities';
 
 // wraps the Targetimage with controls for the various properties
 function Retropixels(props) {
@@ -18,21 +21,43 @@ function Retropixels(props) {
 
   const { jimpImage, filename } = props;
 
+  const [pixelImage, setPixelImage] = useState(undefined);
   const [colorspace, setColorSpace] = useState(colorspaceDefault);
   const [palette, setPalette] = useState(paletteDefault);
   const [hires, setHires] = useState(false);
   const [dither, setDither] = useState(ditherDefault);
   const [ditherRadius, setDitherRadius] = useState(32);
 
-  const parsedFilename = parseFilename(filename);
-  const targetBasename = `${parsedFilename.basename.substring(0, 30)}`;
+  let targetFilename = 'output';
+  if (pixelImage !== undefined) {
+    const extension = pixelImage.mode.pixelWidth < 2 ? '.art' : '.kla';
+    const parsedFilename = parseFilename(filename);
+    targetFilename = `${parsedFilename.basename.substring(0, 30)}${extension}`;
+  }
+
+  function onNewPixelImage(newPixelImage) {
+    setPixelImage(newPixelImage);
+  }
+
+  function saveOutput() {
+    const binary = C64Writer.toBinary(pixelImage);
+    const buffer = C64Writer.toBuffer(binary);
+    const blob = new Blob([buffer], { type: 'application/octet-stream' });
+    saveAs(blob, targetFilename);
+  }
+
+  let outputFormat;
+  if (pixelImage !== undefined) {
+    outputFormat = pixelImage.mode.pixelWidth < 2 ? 'Art studio' : 'Koala';
+  }
 
   return (
     <>
-      <h4>{targetBasename}</h4>
+      <h4>{targetFilename}</h4>
       <Container>
         <TargetImage
           jimpImage={jimpImage}
+          onChanged={onNewPixelImage}
           hires={hires}
           colorspaceId={colorspace}
           paletteId={palette}
@@ -68,7 +93,6 @@ function Retropixels(props) {
           <Grid item>
             <BlurLinearIcon />
           </Grid>
-
           <Grid item xs>
             <Slider
               disabled={dither === 'none'}
@@ -80,6 +104,13 @@ function Retropixels(props) {
             />
           </Grid>
         </Grid>
+      </Container>
+      <Container>
+        <Box m={2}>
+          <Button variant="contained" disabled={pixelImage === undefined} color="primary" onClick={() => saveOutput()}>
+            <CloudDownloadIcon /> &nbsp; Download {outputFormat} file
+          </Button>
+        </Box>
       </Container>
     </>
   );
